competitor_analysis/backend/pyproject.toml
[project]
name = "competitor-analysis"
version = "0.1.0"
description = "Sistema de análise de concorrentes com agentes IA"
authors = [
    {name = "Diogo Gonzaga Ferreira", email = "digadiogoo@gmail.com"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langgraph",
    "langchain-openai",
    "scrapingant-client",
    "python-dotenv",
    "pydantic",
    "httpx",
    "beautifulsoup4",
    "markdown",
    "python-multipart"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8"
]
